requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-trusted-host requirements.in
#
agency-swarm==0.4.4
    # via -r requirements.in
aiofiles==23.2.1
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   gradio
    #   httpx
    #   openai
    #   starlette
argcomplete==3.5.3
    # via datamodel-code-generator
astroid==3.3.8
    # via pylint
black==25.1.0
    # via datamodel-code-generator
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
    #   typer
    #   uvicorn
datamodel-code-generator==0.26.1
    # via agency-swarm
deepdiff==6.7.1
    # via agency-swarm
dill==0.3.9
    # via pylint
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
docstring-parser==0.16
    # via agency-swarm
email-validator==2.2.0
    # via pydantic
fastapi==0.115.8
    # via gradio
ffmpy==0.5.0
    # via gradio
filelock==3.17.0
    # via huggingface-hub
fsspec==2025.2.0
    # via
    #   gradio-client
    #   huggingface-hub
genson==1.3.0
    # via datamodel-code-generator
gradio==5.14.0
    # via -r requirements.in
gradio-client==1.7.0
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   openai
    #   safehttpx
huggingface-hub==0.28.1
    # via
    #   gradio
    #   gradio-client
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
inflect==5.6.2
    # via datamodel-code-generator
isort==5.13.2
    # via
    #   datamodel-code-generator
    #   pylint
jinja2==3.1.5
    # via
    #   datamodel-code-generator
    #   gradio
jiter==0.8.2
    # via openai
jsonref==1.1.0
    # via agency-swarm
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
numpy==2.2.2
    # via
    #   gradio
    #   pandas
openai==1.61.0
    # via agency-swarm
ordered-set==4.1.0
    # via deepdiff
orjson==3.10.15
    # via gradio
packaging==24.2
    # via
    #   black
    #   build
    #   datamodel-code-generator
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.2.3
    # via gradio
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via gradio
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via
    #   black
    #   pylint
pydantic[email]==2.8.2
    # via
    #   agency-swarm
    #   datamodel-code-generator
    #   fastapi
    #   gradio
    #   openai
pydantic-core==2.20.1
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
pylint==3.3.4
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   agency-swarm
python-multipart==0.0.20
    # via gradio
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   datamodel-code-generator
    #   gradio
    #   huggingface-hub
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
rich==13.7.1
    # via
    #   agency-swarm
    #   typer
ruff==0.9.4
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.45.3
    # via
    #   fastapi
    #   gradio
termcolor==2.4.0
    # via agency-swarm
tomlkit==0.13.2
    # via
    #   gradio
    #   pylint
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer==0.15.1
    # via gradio
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2025.1
    # via pandas
unidiff==0.7.5
    # via -r requirements.in
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via gradio
websockets==14.2
    # via gradio-client
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
